docs: rename commands to be consistent with their command
