+++ COMMANDS.md
@@ -1,0 +1,15 @@
+| Command                                                           | Description                                                                                           |
+| ----------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------- |
+| [copy](./src/commands/copy/README.md)                             | Copy a manifest of files                                                                              |
+| [create-manifest](./src/commands/create-manifest/README.md)       | Create a list of files to copy and pass as a manifest                                                 |
+| [group](./src/commands/group/README.md)                           | group a array of inputs into a set                                                                    |
+| [lds-fetch-layer](./src/commands/lds-fetch-layer/README.md)       | Download a LDS layer from the LDS Cache                                                               |
+| [list](./src/commands/list/README.md)                             | List and group files into collections of tasks                                                        |
+| [mapsheet-coverage](./src/commands/mapsheet-coverage/README.md)   | Create a list of mapsheets needing to be created from a basemaps configuration                        |
+| [stac-catalog](./src/commands/stac-catalog/README.md)             | Construct STAC catalog                                                                                |
+| [stac-github-import](./src/commands/stac-github-import/README.md) | Format and push a STAC collection.json file and Argo Workflows parameters file to a GitHub repository |
+| [stac-sync](./src/commands/stac-sync/README.md)                   | Sync STAC files                                                                                       |
+| [stac-validate](./src/commands/stac-validate/README.md)           | Validate STAC files                                                                                   |
+| [tileindex-validate](./src/commands/tileindex-validate/README.md) | List input files and validate there are no duplicates.                                                |
+| [pretty-print](./src/commands/pretty-print/README.md)             | Pretty print JSON files                                                                               |
+| [generate-path](./src/commands/generate-path/README.md)           | Generate target path from collection metadata                                                         |
