+++ src/commands/basemaps-github/make.cog.github.ts
@@ -11,7 +11,7 @@ import { fsa } from '@chunkd/fs';
 import { logger } from '../../log.js';
 import { DEFAULT_PRETTIER_FORMAT } from '../../utils/config.js';
 import { GithubApi } from '../../utils/github.js';
-import { prettyPrint } from '../format/pretty.print.js';
+import { prettyPrint } from '../pretty-print/pretty.print.js';
 
 export enum Category {
   Urban = 'Urban Aerial Photos',
