+++ src/commands/generate-path/README.md
@@ -1,0 +1,28 @@
+# generate-path
+
+Generate target path from collection metadata
+
+## Usage
+
+generate-path <options> <path>
+
+### Arguments
+
+| Usage  | Description                                               | Options |
+| ------ | --------------------------------------------------------- | ------- |
+| <path> | path to source data where collection.json file is located |         |
+
+### Options
+
+| Usage                      | Description                         | Options  |
+| -------------------------- | ----------------------------------- | -------- |
+| --config <str>             | Location of role configuration file | optional |
+| --target-bucket-name <str> | Target bucket name, e.g. nz-imagery |          |
+
+### Flags
+
+| Usage     | Description     | Options |
+| --------- | --------------- | ------- |
+| --verbose | Verbose logging |         |
+
+<!-- This file has been autogenerated by src/readme.generate.ts -->
