+++ src/commands/generate-path/__test__/generate.path.test.ts
