+++ src/commands/generate-path/__test__/path.date.test.ts
