+++ src/commands/generate-path/__test__/sample.ts
