+++ src/commands/generate-path/path.constants.ts
