+++ src/commands/generate-path/path.date.ts
