+++ src/commands/generate-path/path.generate.ts
