+++ src/commands/index.ts
@@ -5,12 +5,12 @@ import { basemapsCreatePullRequest } from './basemaps-github/create-pr.js';
 import { basemapsCreateMapSheet } from './basemaps-mapsheet/create-mapsheet.js';
 import { commandCopy } from './copy/copy.js';
 import { commandCreateManifest } from './create-manifest/create-manifest.js';
-import { commandPrettyPrint } from './format/pretty.print.js';
+import { commandGeneratePath } from './generate-path/path.generate.js';
 import { commandGroup } from './group/group.js';
-import { commandLdsFetch } from './lds-cache/lds.cache.js';
+import { commandLdsFetch } from './lds-fetch-layer/lds.fetch.layer.js';
 import { commandList } from './list/list.js';
 import { commandMapSheetCoverage } from './mapsheet-coverage/mapsheet.coverage.js';
-import { commandGeneratePath } from './path/path.generate.js';
+import { commandPrettyPrint } from './pretty-print/pretty.print.js';
 import { commandStacCatalog } from './stac-catalog/stac.catalog.js';
 import { commandStacGithubImport } from './stac-github-import/stac.github.import.js';
 import { commandStacSync } from './stac-sync/stac.sync.js';
