+++ src/commands/lds-fetch-layer/README.md
@@ -1,0 +1,28 @@
+# lds-fetch-layer
+
+Download a LDS layer from the LDS Cache
+
+## Usage
+
+lds-fetch-layer <options> [...str]
+
+### Arguments
+
+| Usage    | Description                                   | Options |
+| -------- | --------------------------------------------- | ------- |
+| [...str] | Layer id and optional version "layer@version" |         |
+
+### Options
+
+| Usage          | Description                         | Options  |
+| -------------- | ----------------------------------- | -------- |
+| --config <str> | Location of role configuration file | optional |
+| --target <str> | Target directory to save files      |          |
+
+### Flags
+
+| Usage     | Description     | Options |
+| --------- | --------------- | ------- |
+| --verbose | Verbose logging |         |
+
+<!-- This file has been autogenerated by src/readme.generate.ts -->
