+++ src/commands/pretty-print/README.md
@@ -1,0 +1,28 @@
+# pretty-print
+
+Pretty print JSON files
+
+## Usage
+
+pretty-print <options> <path>
+
+### Arguments
+
+| Usage  | Description                       | Options |
+| ------ | --------------------------------- | ------- |
+| <path> | Path of the files to pretty print |         |
+
+### Options
+
+| Usage          | Description                                                                              | Options  |
+| -------------- | ---------------------------------------------------------------------------------------- | -------- |
+| --config <str> | Location of role configuration file                                                      | optional |
+| --target <str> | Use if files have to be saved somewhere else instead of overwriting the source (testing) | optional |
+
+### Flags
+
+| Usage     | Description     | Options |
+| --------- | --------------- | ------- |
+| --verbose | Verbose logging |         |
+
+<!-- This file has been autogenerated by src/readme.generate.ts -->
